admin directory delete
