+++ internal/registrar.go
@@ -1,31 +1,40 @@
-package internal
+package admin
 
 import (
 	"encoding/json"
 	"fmt"
 	"log"
+	"oops/main/internal"
+
+	//"oops/main/analytics"
+	//"oops/main/domain"
 	"os"
 )
 
 type Registrar struct {
-	students    []Student
-	courses     []Course
-	enrollments []Enrollment
+	students    []internal.Student
+	courses     []internal.Course
+	enrollments []internal.Enrollment
+	graders     map[int]internal.Grader
 }
 
-func (r *Registrar) AddStudent(s Student) {
+func NewRegistrar() *Registrar {
+	return &Registrar{graders: make(map[int]internal.Grader)}
+}
+
+func (r *Registrar) AddStudent(s internal.Student) {
 	r.students = append(r.students, s)
 }
 
-func (r *Registrar) AddCourse(c Course) {
+func (r *Registrar) AddCourse(c internal.Course) {
 	r.courses = append(r.courses, c)
 }
 
-func (r *Registrar) Enroll(e Enrollment) {
+func (r *Registrar) Enroll(e internal.Enrollment) {
 	r.enrollments = append(r.enrollments, e)
 }
 
-func (r *Registrar) SetGrader(courseID int, g Grader) {
+func (r *Registrar) SetGrader(courseID int, g internal.Grader) {
 	for i, e := range r.enrollments {
 		if e.Course.Id == courseID {
 			r.enrollments[i].Grader = g
@@ -33,7 +42,7 @@ func (r *Registrar) SetGrader(courseID int, g Grader) {
 	}
 }
 
-func (r *Registrar) Enrollments() []Enrollment {
+func (r *Registrar) Enrollments() []internal.Enrollment {
 	return r.enrollments
 }
 
@@ -59,7 +68,7 @@ func (regis *Registrar) LoadCourses() {
 		log.Fatal("Failed to unmarshal courses:", err)
 	}
 	for _, course := range coursesData {
-		regis.AddCourse(NewCourse(course.ID, course.Name))
+		regis.AddCourse(internal.NewCourse(course.ID, course.Name))
 	}
 }
 
@@ -74,9 +83,9 @@ func (r *Registrar) LoadStudents() {
 	if err != nil {
 		log.Fatal("Failed to unmarshal students:", err)
 	}
-	r.students = make([]Student, 0, len(students))
+	r.students = make([]internal.Student, 0, len(students))
 	for _, sd := range students {
-		student := NewStudent(sd.ID, sd.Name)
+		student := internal.NewStudent(sd.ID, sd.Name)
 		r.AddStudent(student)
 	}
 }
@@ -88,7 +97,8 @@ func (r *Registrar) DisplayStudents() {
 }
 
 func (r *Registrar) DisplayCourses() {
+	log.Println("Courses...")
 	for _, cr := range r.courses {
-		fmt.Printf("#%d : %s\n", cr.Id, cr.Name)
+		log.Printf("#%d : %s\n", cr.Id, cr.Name)
 	}
 }
